+++ R/tools.R
@@ -1,7 +1,16 @@
-set_server_tools <- function(x, x_arg = caller_arg(x), call = caller_env()) {
+set_server_tools <- function(
+  x,
+  session_tools = TRUE,
+  x_arg = caller_arg(x),
+  call = caller_env()
+) {
   if (is.null(x)) {
-    the$server_tools <- c(list(list_r_sessions_tool, select_r_session_tool))
-    return()
+    if (session_tools) {
+      the$server_tools <- c(list(list_r_sessions_tool, select_r_session_tool))
+      return()
+    } else {
+      cli::cli_abort("No tools selected to serve.", call = call)
+    }
   }
 
   # evaluate eagerly so that caller arg is correct if `looks_like_r_file()`
@@ -22,7 +31,11 @@ set_server_tools <- function(x, x_arg = caller_arg(x), call = caller_env()) {
     )
   }
 
-  if (!is_list(x) || !all(vapply(x, inherits, logical(1), "ellmer::ToolDef"))) {
+  if (!is.list(x)) {
+    x <- list(x)
+  }
+
+  if (!all(vapply(x, inherits, logical(1), "ellmer::ToolDef"))) {
     msg <-
       "{.arg {x_arg}} must be a list of tools created with {.fn ellmer::tool}
        or a .R file path that returns a list of ellmer tools when sourced."
@@ -39,19 +52,22 @@ set_server_tools <- function(x, x_arg = caller_arg(x), call = caller_env()) {
     )
   ) {
     cli::cli_abort(
-      "The tool names {.field list_r_sessions} and {.field select_r_session} are 
+      "The tool names {.field list_r_sessions} and {.field select_r_session} are
        reserved by {.pkg mcptools}.",
       call = call
     )
   }
 
-  the$server_tools <- c(
-    x,
-    list(
-      list_r_sessions_tool,
-      select_r_session_tool
+  if (session_tools) {
+    x <- c(
+      x,
+      list(
+        list_r_sessions_tool,
+        select_r_session_tool
+      )
     )
-  )
+  }
+  the$server_tools <- x
 }
 
 looks_like_r_file <- function(x) {
@@ -109,18 +125,10 @@ list_r_sessions_description <- paste(
   "and call select_r_session unless the user asks you to specifically."
 )
 
-list_r_sessions_tool <-
-  if (is_new_ellmer()) {
-    ellmer::tool(
-      fun = list_r_sessions,
-      description = list_r_sessions_description
-    )
-  } else {
-    ellmer::tool(
-      .fun = list_r_sessions,
-      .description = list_r_sessions_description
-    )
-  }
+list_r_sessions_tool <- ellmer::tool(
+  fun = list_r_sessions,
+  description = list_r_sessions_description
+)
 
 select_r_session <- function(session) {
   nanonext::reap(the$server_socket[["dialer"]][[1L]])
@@ -145,22 +153,13 @@ select_r_session_description <- paste(
   "call this tool more than once if you need to switch between sessions."
 )
 
-select_r_session_tool <-
-  if (is_new_ellmer()) {
-    ellmer::tool(
-      fun = select_r_session,
-      description = select_r_session_description,
-      arguments = list(
-        session = ellmer::type_integer("The R session number to select.")
-      )
-    )
-  } else {
-    ellmer::tool(
-      .fun = select_r_session,
-      .description = select_r_session_description,
-      session = ellmer::type_integer("The R session number to select.")
-    )
-  }
+select_r_session_tool <- ellmer::tool(
+  fun = select_r_session,
+  description = select_r_session_description,
+  arguments = list(
+    session = ellmer::type_integer("The R session number to select.")
+  )
+)
 
 get_mcptools_tools <- function() {
   # must be called inside of the server session